Implemented support for the MUMPS LU solver
MUMPS (MUltifrontal Massively Parallel sparse direct Solver) is a
parallel direct linear solver for sparse matrices and is considered to
be one of the better solvers available. PETSc has built-in support for
using the MUMPS solver and in this commit we add support to DREAM for
running with the MUMPS solver instead of the standard PETSc LU or GMRES
solvers.

Initial tests suggest a ~35% increase in performance with MUMPS compared
to the standard PETSc LU solver. CPU core utilization ranged between
35-50% (of 4 CPU cores) for MUMPS, and between 20-25% for the standard
solver (i.e. single core usage). We note however that the test (a highly
resolved run of the runaway example) was perhaps not optimal for
benchmarking the solvers, as much time was still spent in constructing
the matrices, but should serve as a decent indication nonetheless.