Reimplemented access to individual equation terms
This commit contains a new way of accessing individual equation terms
from the ``OtherQuantityHandler``. The new method does not rely on term
IDs and does not require additional memory to be allocated (aside from
the memory already allocated by ``QuantityData`` when storing the data
for the ``OtherQuantity``). Instead, we keep track of interesting terms
in a separate struct which is passed around in ``SimulationGenerator``
and finally ends up with the ``OtherQuantityHandler``.

I have not been able to verify that this commit works yet as there
appears to be a bug on master which causes all suitable examples to
crash.