Fixed inconsistent NNZ in DiffusionTerm
When disabling the hot-tail grid in the runaway example, DREAM would
crash with PETSc complaining about some row getting more non-zero
elements than total elements. This turned out to be due to the number of
nonzeros for the DiffusionTerm being hardcoded to 11, which is the
_maximum_ number of non-zeros which can be set for that term and only
occurs when the diffusion is applied to a kinetic grid. This fix ensures
that only a realistic number of non-zeros are set depending on the type
of grid used.