Fixed prefactors in HeatTransport RR
The diffusion coefficient now seems to take reasonable values, but it
also seems to be highly dependent on dt (and independent of tMax). The
transport-rr example yields what appears to be a steady-state solution
after the first time step, which then remains steady throughout the
remainder of the simulation with only sub-tolerance variations.