Added the SvenssonTransport terms to the Transport setting.
On branch SvenssonTransport
Your branch is ahead of 'origin/SvenssonTransport' by 1 commit.
  (use "git push" to publish your local commits)

Changes to be committed:
	modified:   include/DREAM/Equations/Fluid/SvenssonTransport.hpp
	modified:   include/DREAM/Equations/Fluid/SvenssonTransport.tcc
	modified:   include/DREAM/Settings/SimulationGenerator.hpp
	modified:   src/Equations/Fluid/SvenssonTransport.cpp
	modified:   src/Settings/Equations/E_field.cpp
	modified:   src/Settings/Equations/T_cold.cpp
	modified:   src/Settings/Equations/distribution.cpp
	modified:   src/Settings/Equations/n_re.cpp
	modified:   src/Settings/Transport.cpp

Due to the requirements SvenssonTransport to have the `EquationSystem`
variables `unknonwns` and `REFluid`, the structure of the
`ConstructTransportTerm` constructor in `Transport.cpp` had to be
modified, and all subsequent calls to it, to also include an
`EquationSystem*` input argument.

I have also created a `ConstructSvenssonTransportTerm_internal` in
`Transport.cpp` for constructing the three SvenssonTranport terms.