Added capabilities for setting `pstar` from the Python init file.
Changes to be committed:
	modified:   examples/SvenssonTransport/generate.py
	modified:   include/DREAM/Equations/Fluid/SvenssonTransport.tcc
	modified:   py/DREAM/Settings/TransportSettings.py
	modified:   src/Settings/Transport.cpp

At the moment `pstar` is being set separately for `s_ar` and
`s_drr`. I'm not sure if this is bad or not. It's not as intended in
the `Transport.cpp` file, where I've now defined the same setting for
both `s_ar` and `s_drr`. Perhpas it's good to have that capability in
the future.

I also fix (I think) some bugs in `fromdict` regarding how the `s_ar`
settings are being read. There is still bugs for the "regular" `ar`
settings (I think).

For now, if `pstar` is not given, it's set to 0, but perhaps it's
better to raise an exception instead.