Implemented consistent initialization of f_re
When prescribing a runaway electron density profile in fluid-kinetic
mode we would like the runaway distribution f_re to be initialized
consistently with n_re. In this commit we do so by putting all particles
of n_re in the p=0, xi=+1 grid cell on f_re. This should be consistent
with the reasoning we use for adding fluid runaway source terms in the
same cell to the runaway grid in fluid-kinetic mode.

One case not handled in the current implementation is if E<0 at the
beginning of the simulation, in which case the particles should rather
be put in xi=-1.