Refactored linear solver code. Updated docs.
The 'SolverLinearlyImplicit' and 'SolverNonLinear' solvers have until
now had separate implementations for creating linear solvers. With this
commit I move that code to a new 'SelectLinearSolver()' routine in
'Solver.cpp' so that we don't have to implement new linear solvers in
more than one place.

I also update the information about compiling DREAM in the
documentation. The Sphinx documentation is now the most complete
description on how to compile DREAM and PETSc (previously it was the
README).