Minor breaking change to SetJacobianBlock()
This applies to all EquationTerms and BoundaryConditions. The method
SetJacobianBlock() should now return a boolean indicating whether or not
any jacobian elements were actually set in the call. This can help in
debugging which terms actually contribute to a given element of the
jacobian. In the future, it could perhaps also be used to optimize
construction of the jacobian.