Removed use of 'sincos()' function
This function is not part of the C++ standard and only exists in
platform-specific math libraries. This causes problems when compiling on
platforms other than Linux. The compiler should be able to optimize the
consecutive calls to 'sin()' and 'cos()' when optimizations are enabled
(which should be the default case anyway).