Added extremum check to numeric magnetic field
The bounce-averaged theory of DREAM is designed for magnetic fields with
exactly one maximum and one minimum in the magnetic field strength along
a given magnetic field line. In case more maxima/minima occur along a
given magnetic field line, the theory breaks down and DREAM should
indicate this to the user.