Implement frozen current mode
The frozen current mode adjusts the radial transport such that the total
plasma current Ip always matches a prescribed value. In this commit, the
frozen current mode works for the linearly implicit solver, but remains
numerically unstable in the non-linear solver. This will next be fixed
by integrating the (new) external iterator with the non-linear solver as
well.